+++ labyrinth/src/consts.rs
@@ -1,3 +1,1 @@
-pub const PLAYER_SPEED: f32 = 250.0;
-
 pub const SPRITE_SIZE: f32 = 64.0;