+++ labyrinth/src/game/player/systems.rs
@@ -1,14 +1,39 @@
+use std::time::Duration;
+
 use bevy::prelude::*;
 use bevy::sprite::collide_aabb::collide;
 
+use super::super::components::*;
+use super::PlayerState;
+use crate::game::events::MoveRequested;
 use crate::AppState;
-use crate::{consts::*, game::events::MoveRequested};
 
-use super::super::components::{ExitDoorCollider, Player, WallCollider};
+pub fn player_is_on_the_move(current_player_state: Res<State<PlayerState>>) {
+    if current_player_state.0 == PlayerState::Moving {
+        println!("Moving")
+    }
+}
+
+pub fn player_is_moving(
+    mut commands: Commands,
+    player_query: Query<(Entity, &MovingPlayer), With<MovingPlayer>>,
+    time: Res<Time>,
+) {
+    for (entity, movement) in player_query.iter() {
+        if let Some(last_updated) = time.last_update() {
+            let elapsed = last_updated - movement.started;
+            if elapsed > Duration::from_millis(150) {
+                commands.entity(entity).remove::<MovingPlayer>();
+            }
+        }
+    }
+}
 
 pub fn player_movement(
-    mut player_query: Query<&mut Transform, With<Player>>,
+    mut commands: Commands,
+    mut player_query: Query<&mut Transform, (With<Player>, Without<MovingPlayer>)>,
     mut camera_query: Query<&mut Transform, (With<Camera2d>, Without<Player>)>,
+    player_entity: Query<Entity, With<Player>>,
     wall_collider_query: Query<
         &Transform,
         (With<WallCollider>, Without<Player>, Without<Camera2d>),
@@ -18,19 +43,23 @@ pub fn player_movement(
         (With<ExitDoorCollider>, Without<Player>, Without<Camera2d>),
     >,
     mut events: EventReader<MoveRequested>,
-    //    keyboard_input: Res<Input<KeyCode>>,
     mut next_app_state: ResMut<NextState<AppState>>,
     time: Res<Time>,
 ) {
-    let time_delta = time.delta_seconds();
     if let Ok(mut transform) = player_query.get_single_mut() {
         for event in events.iter() {
             let step_x = Vec3::new(event.direction.x, 0.0, 0.0);
             let step_y = Vec3::new(0.0, event.direction.y, 0.0);
 
-            let step_x = step_x * PLAYER_SPEED * time_delta;
-            let step_y = step_y * PLAYER_SPEED * time_delta;
-
+            let step_x = step_x * 64.0; //PLAYER_SPEED * time_delta;
+            let step_y = step_y * 64.0; //PLAYER_SPEED * time_delta;
+            if let Ok(player_entity) = player_entity.get_single() {
+                if let Some(last_updated) = time.last_update() {
+                    commands.entity(player_entity).insert(MovingPlayer {
+                        started: last_updated,
+                    });
+                }
+            }
             #[allow(clippy::clone_on_copy)]
             let mut target = transform.translation.clone();