+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - [I changed to event-based control](devblog/2023-04-21.md#event-based-control)
 - [Potential mobile web play](devblog/2023-04-21.md#mobile-play)
+- [Moving tile by tile](devblog/2023-04-21.md#moving-tile-by-tile)
 
 ## 2023-04-20